refactor(qwik/a2ui): consolidate ElmA2uiSurfaceRenderer functionality into ElmA2uiRenderer and remove unused surface renderer files
