+++ packages/qwik/src/components/a2ui/elm-a2ui-catalog-renderer.stories.tsx
@@ -18,7 +18,7 @@ import {
 } from "@a2ui/web_core/v0_9/basic_catalog";
 import { type CatalogRendererMap } from "./elm-a2ui-catalog-renderer";
 import { elmBasicCatalogRendererMap } from "./elm-a2ui-basic-catalog-renderer";
-import { renderTree, findRootId } from "./elm-a2ui-surface-renderer";
+import { renderTree, findRootId } from "./elm-a2ui-renderer";
 
 const CATALOG_ID = "https://a2ui.org/specification/v0_9/basic_catalog.json";
